fix: lazily load market data for fiat ramp assets (#3340)
