+++ src/state/apis/fiatRamps/selectors.ts
@@ -5,23 +5,8 @@ import { createSelector } from 'reselect'
 import { createDeepEqualOutputSelector } from 'state/selector-utils'
 import { defaultMarketData } from 'state/slices/marketDataSlice/marketDataSlice'
 import { selectAssets, selectMarketData } from 'state/slices/selectors'
-import { useAppSelector } from 'state/store'
 
 import { fiatRampApi } from './fiatRamps'
-
-export const useIsSupportedFiatBuyAssetId = (assetId: AssetId): boolean =>
-  Boolean(
-    (useAppSelector(fiatRampApi.endpoints.getFiatRamps.select()).data?.buyAssetIds ?? []).includes(
-      assetId,
-    ),
-  )
-
-export const useIsSupportedFiatSellAssetId = (assetId: AssetId): boolean =>
-  Boolean(
-    (useAppSelector(fiatRampApi.endpoints.getFiatRamps.select()).data?.sellAssetIds ?? []).includes(
-      assetId,
-    ),
-  )
 
 export const selectFiatBuyAssetIds = createDeepEqualOutputSelector(
   fiatRampApi.endpoints.getFiatRamps.select(),
